Removed cryptography kotlin logic from encryption content algorithm
